requirements.txt
oauthlib
requests-oauthlib
requests
pyyaml